pyproject.toml
[tool.poetry]
name = "sudocv"
version = "0.1.0"
description = ""
authors = ["John Aldrich Bernardo <4ldrich@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
pygithub = "^1.45"
python-docx = "^0.8.10"
click = "^7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
